Extract more options provider
