+++ arangod/Cluster/ReplicationTimeoutOptionsProvider.cpp
@@ -1,0 +1,118 @@
+////////////////////////////////////////////////////////////////////////////////
+/// DISCLAIMER
+///
+/// Copyright 2014-2026 ArangoDB GmbH, Cologne, Germany
+/// Copyright 2004-2014 triAGENS GmbH, Cologne, Germany
+///
+/// Licensed under the Business Source License 1.1 (the "License");
+/// you may not use this file except in compliance with the License.
+/// You may obtain a copy of the License at
+///
+///     https://github.com/arangodb/arangodb/blob/devel/LICENSE
+///
+/// Unless required by applicable law or agreed to in writing, software
+/// distributed under the License is distributed on an "AS IS" BASIS,
+/// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+/// See the License for the specific language governing permissions and
+/// limitations under the License.
+///
+/// Copyright holder is ArangoDB GmbH, Cologne, Germany
+///
+////////////////////////////////////////////////////////////////////////////////
+
+#include "ReplicationTimeoutOptionsProvider.h"
+
+#include "Logger/LogMacros.h"
+#include "Logger/LogTopic.h"
+#include "ProgramOptions/Parameters.h"
+#include "ProgramOptions/ProgramOptions.h"
+
+namespace arangodb::replication_timeout {
+
+using namespace arangodb::options;
+
+void ReplicationTimeoutOptionsProvider::declareOptions(
+    std::shared_ptr<ProgramOptions> options,
+    ReplicationTimeoutFeatureOptions& opts) {
+  options
+      ->addOption("--cluster.synchronous-replication-timeout-minimum",
+                  "All synchronous replication timeouts are at least "
+                  "this value (in seconds).",
+                  new DoubleParameter(&opts.lowerLimit),
+                  makeDefaultFlags(Flags::Uncommon, Flags::OnDBServer))
+      .setLongDescription(R"(**Warning**: This option should generally remain
+untouched and only be changed with great care!
+
+The minimum timeout in seconds for the internal synchronous replication
+mechanism between DB-Servers. If replication requests are slow, but the servers
+are otherwise healthy, timeouts can cause followers to be dropped unnecessarily,
+resulting in costly resync operations. Increasing this value may help avoid such
+resyncs. Conversely, decreasing it may cause more resyncs, while lowering the
+latency of individual write operations.
+
+**Warning**: If you use multiple DB-Servers, use the same value on all
+DB-Servers.)");
+
+  options
+      ->addOption("--cluster.synchronous-replication-timeout-maximum",
+                  "All synchronous replication timeouts are at most "
+                  "this value (in seconds).",
+                  new DoubleParameter(&opts.upperLimit),
+                  makeDefaultFlags(Flags::Uncommon, Flags::OnDBServer))
+      .setIntroducedIn(30800)
+      .setLongDescription(R"(**Warning**: This option should generally remain
+untouched and only be changed with great care!
+
+Extend or shorten the timeouts for the internal synchronous replication
+mechanism between DB-Servers. All such timeouts are affected by this change.
+
+**Warning**: If you use multiple DB-Servers, use the same value on all
+DB-Servers.)");
+
+  options
+      ->addOption(
+          "--cluster.synchronous-replication-timeout-factor",
+          "All synchronous replication timeouts are multiplied by this factor.",
+          new DoubleParameter(&opts.timeoutFactor),
+          makeDefaultFlags(Flags::Uncommon, Flags::OnDBServer))
+      .setLongDescription(R"(**Warning**: If you use multiple DB-Servers, use
+the same value on all DB-Servers.)");
+
+  options
+      ->addOption("--cluster.synchronous-replication-timeout-per-4k",
+                  "All synchronous replication timeouts are increased by this "
+                  "amount per 4096 bytes (in seconds).",
+                  new DoubleParameter(&opts.timeoutPer4k),
+                  makeDefaultFlags(Flags::Uncommon, Flags::OnDBServer))
+      .setLongDescription(R"(**Warning**: If you use multiple DB-Servers, use
+the same value on all DB-Servers.)");
+
+  options
+      ->addOption(
+          "--cluster.shard-synchronization-attempt-timeout",
+          "The timeout (in seconds) for every shard synchronization attempt. "
+          "Running into the timeout does not lead to a synchronization "
+          "failure, but continues the synchronization shortly after. "
+          "Setting a timeout can help to split the replication of large "
+          "shards into smaller chunks and release snapshots on the "
+          "leader earlier.",
+          new DoubleParameter(&opts.shardSynchronizationAttemptTimeout),
+          makeDefaultFlags(Flags::Uncommon, Flags::OnDBServer))
+      .setIntroducedIn(30902)
+      .setLongDescription(R"(**Warning**: If you use multiple DB-Servers, use
+the same value on all DB-Servers.)");
+}
+
+void ReplicationTimeoutOptionsProvider::validateOptions(
+    std::shared_ptr<ProgramOptions> /*options*/,
+    ReplicationTimeoutFeatureOptions& opts) {
+  if (opts.upperLimit < opts.lowerLimit) {
+    LOG_TOPIC("8a9f3", WARN, Logger::CONFIG)
+        << "--cluster.synchronous-replication-timeout-maximum must be at least "
+        << "--cluster.synchronous-replication-timeout-minimum, setting max to "
+        << "min";
+    opts.upperLimit = opts.lowerLimit;
+  }
+}
+
+}  // namespace arangodb::replication_timeout
